+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added reset event log button (#9, @ilyaryabchinski)
 - Added status bar with overall event statistics
 - Fixed event loading throttling delay which was committed by mistake
+- Fixed various bugs with event log displaying
 
 ## 0.1.2 (August 25, 2021)